requirements.txt
discord.py[voice]>=2.4.0
git+https://github.com/JordanLevy99/discord-ext-voice-recv.git
python-dotenv
pandas
pydrive
APScheduler
tabulate
dropbox
PyNaCl
pytz
openai
mutagen
finnhub-python
yt-dlp
newsapi-python
pydub
webrtcvad
websockets
numpy
Flask>=2.0.0
Werkzeug>=2.0.0